flights_agent/requirements.txt
# NL-to-SQL Agent — pinned dependencies
# Install with: pip install -r requirements.txt

# LangChain orchestration
langchain==0.3.7
langchain-anthropic==0.3.3
langchain-community==0.3.7
# langchain-core: intentionally unpinned — version is resolved by langchain + langchain-anthropic

# Claude API: intentionally unpinned — version is resolved by langchain-anthropic
# (langchain-anthropic 0.3.3 requires anthropic>=0.41.0,<1)

# Vector store
chromadb==0.5.23
sentence-transformers==3.3.1
onnxruntime==1.17.1
# torch is installed separately in Dockerfile (CPU-only build) to prevent GPU version being pulled in

# BigQuery
google-cloud-bigquery==3.27.0
db-dtypes==1.3.1
pyarrow==18.1.0

# Data
pandas==2.2.3

# Frontend
streamlit==1.40.0
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
google-cloud-secret-manager
